[GAFF] Added AlignedMalloc() and AlignedFree() functions. [SHIBBOLETH] Added IAllocator. Moved all allocation functions to Memory*.dll. Added premake options for Link-Time Code Generation and Symbol information generation. Fixed debug code in Allocator. Mostly added allocator debug feature for capturing stacktrace information for allocations.

+++ src/Game/Shibboleth_SetupGraphicsState.cpp
@@ -234,7 +234,7 @@ void SetupGraphicsState::update(void)
 		}
 
 		if (icon) {
-			if (!render_manager.getWindowData(render_manager.getNumWindows() - 1).window->setIcon(icon.getString())) {
+			if (!render_manager.getWindowData(static_cast<unsigned int>(render_manager.getNumWindows() - 1)).window->setIcon(icon.getString())) {
 				log.first.printf("WARNING - Failed to set window icon to '%s'.\n", icon.getString());
 			}
 		}
